Move back pb to internal; Organize modules
