+++ back/gengo
@@ -1,0 +1,34 @@
+#!/usr/bin/env bash
+
+function checkDepOr() {
+    ! (hash "${1}" 2>/dev/null || hash "${2}" 2>/dev/null) && echerr "requires ${1} or ${2}"
+}
+
+checkDepOr "realpath" "grealpath"
+
+function depOr() {
+    local ret="${1}"
+    hash "${2}" 2>/dev/null && ret="${2}"
+    echo "${ret}"
+}
+
+function xrealpath() {
+    local cmd="$(depOr "realpath" "grealpath")"
+    ${cmd} "${@}"
+}
+
+function main() {
+	local pkgs=(\
+		"internal/pb" \
+		"internal/httpsrv" \
+	)
+    	local dir="$(xrealpath "${0}" | xargs dirname)"
+
+	for pkg in "${pkgs[@]}"; do
+		pushd "${dir}/${pkg}"
+		go generate
+		popd
+	done
+}
+
+main
