--- back/gogen.go
@@ -1,4 +1,0 @@
-package back
-
-//go:generate go generate ./pb
-//go:generate go generate ./internal/httpsrv
