+++ back/internal/go.mod
@@ -1,17 +1,15 @@
-module github.com/champagneabuelo/openboard/back
+module github.com/champagneabuelo/openboard/back/internal
 
 go 1.12
 
 require (
-	github.com/codemodus/alfred v0.2.1
 	github.com/codemodus/chain/v2 v2.1.2
 	github.com/codemodus/hedrs v0.1.1
 	github.com/codemodus/mixmux v0.2.1
-	github.com/codemodus/sigmon/v2 v2.0.1
 	github.com/codemodus/swagui v0.2.2
-	github.com/golang/protobuf v1.3.0
-	github.com/grpc-ecosystem/grpc-gateway v1.7.0
+	github.com/golang/protobuf v1.2.0
+	github.com/grpc-ecosystem/grpc-gateway v1.8.3
 	golang.org/x/net v0.0.0-20190213061140-3a22650c66bd
 	google.golang.org/genproto v0.0.0-20190219182410-082222b4a5c5
-	google.golang.org/grpc v1.18.0
+	google.golang.org/grpc v1.19.0
 )
