+++ back/internal/httpsrv/httpsrv.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/champagneabuelo/openboard/back/internal/httpsrv/internal/embed/swagger"
-	"github.com/champagneabuelo/openboard/back/pb"
+	"github.com/champagneabuelo/openboard/back/internal/pb"
 	"github.com/codemodus/chain/v2"
 	"github.com/codemodus/hedrs"
 	"github.com/codemodus/mixmux"
